+++ site/docs/pages/api/types.mdx
@@ -40,6 +40,12 @@ type BuildSwapTransactionParams = GetSwapQuoteParams & {
 };
 ```
 
+## `BuildSwapTransactionResponse`
+
+```ts
+type BuildSwapTransactionResponse = BuildSwapTransaction | APIError;
+```
+
 ## `GetSwapQuoteParams`
 
 ```ts
