+++ src/api/index.ts
@@ -7,7 +7,9 @@ export type {
   APIError,
   BuildPayTransactionParams,
   BuildPayTransactionResponse,
+  BuildSwapTransaction,
   BuildSwapTransactionParams,
+  BuildSwapTransactionResponse,
   GetSwapQuoteParams,
   GetSwapQuoteResponse,
   GetTokensOptions,
